+++ History.md
@@ -3,6 +3,7 @@
 ## New and Noteworthy
 
 ## Enhancements
+* [#120](https://github.com/pmd/pmd-regression-tester/pull/120): Support new PMD 7 binary dist filename
 
 ## Fixed Issues